Checking API's and Fixing any issues
