+++ src/Model/Account/CreateUpdateAccountRequestModel.cs
@@ -1,14 +1,17 @@
 ﻿namespace Tutkoo.mintyfusion.Moov.Sdk.Model.Account
 {
-    #region Namespace
+    #region namespace
     using System.Text.Json.Serialization;
-    #endregion
+    #endregion namespace
 
     #region Class
     public class CreateUpdateAccountRequestModel : AccountRequestModel
     {
+        #region Properties
+        /// <summary>Use AccountType.Value() to se this Value.</summary>
         [JsonPropertyName("accountType")]
         public string AccountType { get; set; }
+        #endregion Properties
     }
     #endregion Class
 }
